+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog 1.0.0].
 
+## Unreleased
+
+### Feat
+
+- **Document**: add method to save document to MarkLogic
+
+### Refactor
+
+- swap from defusedxml to lxml for serialisation
+
 ## v45.0.1 (2026-04-28)
 
 ### Fix